fix for old compilers

+++ crc32.c
@@ -441,7 +441,7 @@ local void write_table64(FILE *out, const z_word_t FAR *table, int k) {
 
     for (n = 0; n < k; n++)
         fprintf(out, "%s0x%016llx%s", n == 0 || n % 3 ? "" : "    ",
-                (unsigned long long)(table[n]),
+                (unsigned __int64)(table[n]),
                 n == k - 1 ? "" : (n % 3 == 2 ? ",\n" : ", "));
 }
